requirements-dev.txt
pytest
tiktoken
requests
matplotlib
wandb
runpod
datasets
random-fourier-features-pytorch
num2words
sympy

# Torch comes with the runpod image